Renamed server.js to app.js and created a server folder to put server related files
